requirements.txt
torch==2.4.0
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx
pandas==2.1.4
aiofiles==23.2.1
fastapi==0.114.2
ffmpy==0.4.0
gradio==4.44.0
gradio-client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
orjson==3.10.7
pydub==0.25.1
python-multipart==0.0.9
ruff==0.6.5
semantic-version==2.10.0
starlette==0.38.5
tomlkit==0.12.0
uvicorn==0.30.6
websockets==12.0
types-requests==2.32.0.20240914
pinecone-plugin-interface==0.0.7
parameterized==0.9.0
jmespath==1.0.1
httpx-sse==0.4.0
fastavro==1.9.7
pinecone-plugin-inference==1.1.0
botocore==1.35.19
s3transfer==0.10.2
pinecone-client==5.0.1
boto3==1.35.19
sentence-transformers==3.1.0
cohere==5.9.2
python-dotenv
